requirements.txt
streamlit==1.31.0
requests==2.32.5
plotly==5.18.0
pandas==2.2.0
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.27.0
openai==1.54.0
pydantic==2.5.3
sqlalchemy==2.0.25
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
httpx==0.27.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.0
